Added Embedding layer to beginning of model
This changes the input format from embeddings (arrays of word vectors)
to word indices. The main motivation for this is to greatly simplify
debugging; It's now possible to obtain the original sentence from the
array of word indices.

Another interesting side-effect is that the word embeddings can be
fine-tuned during training. This adds training time but should
hopefully make for a more powerful model.